Discord Banner, and cleanup of reboot command
