+++ decksite/controllers/api.py
@@ -1,7 +1,7 @@
 import datetime
 import json
 import sys
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, cast
+from typing import Any, Dict, List, Optional, Tuple, cast
 
 from flask import Response, request, session, url_for
 from flask_restx import Resource, fields
