+++ discordbot/commands/restartbot.py
@@ -9,5 +9,10 @@ from shared import redis_wrapper
 # @check(is_owner())
 async def restartbot(ctx: MtgContext) -> None:
     """Restart the bot."""
+    if redis_wrapper.get_bool('discordbot:do_reboot'):
+        await ctx.send('rebooting!')
+        await ctx.bot.stop()
+        return
+
     await ctx.send('Scheduling reboot')
     redis_wrapper.store('discordbot:do_reboot', True)
