+++ magic/image_fetcher.py
@@ -203,3 +203,46 @@ async def generate_banner(names: List[str], background: str, v_crop: int = None)
 
     canvas.save(out_filepath)
     return out_filepath
+
+async def generate_discord_banner(names: List[str], background: str, v_crop: int = None) -> str:
+    cards = [oracle.load_card(name) for name in names]
+    hq_artcrops = fetcher.hq_artcrops()
+    hq = False
+    if background in hq_artcrops.keys():
+        hq = True
+        if v_crop is None:
+            v_crop = hq_artcrops[background][1]
+
+    if v_crop is None:
+        v_crop = 33
+
+    out_filepath = determine_filepath(cards, f'discord-banner-{background}{"hq" if hq else ""}{v_crop}-', '.png')
+
+    if fetch_tools.acceptable_file(out_filepath):
+        return out_filepath
+
+    canvas = Image.new('RGB', (1920, 1080))
+    c = oracle.load_card(background)
+    file_path = await download_art_crop(c, hq_artcrops)
+    if file_path:
+        with Image.open(file_path) as img:
+            canvas.paste(img.resize((1920, 1080), Image.BICUBIC).crop((0, 0, 1920, 1080)))
+
+    n = math.ceil(len(cards) / 2)
+    x = 200
+    for c in cards[:n]:
+        ip = await download_scryfall_png(c)
+        with Image.open(ip) as img:
+            img = img.resize((320, 426), Image.LANCZOS)
+            canvas.paste(img, (x, 100))
+            x = x + img.width + 10
+    x = 300
+    for c in cards[n:]:
+        ip = await download_scryfall_png(c)
+        with Image.open(ip) as img:
+            img = img.resize((320, 426), Image.LANCZOS)
+            canvas.paste(img, (x, 200))
+            x = x + img.width + 10
+
+    canvas.save(out_filepath)
+    return out_filepath
